+++ Pysrcs/tools.py
@@ -11,7 +11,7 @@ def kill_jobs(programList):
 	for program in programList:
 		if program.state != "Finished" and program.state != "Not started":
 			for pid in program.pidList:
-				if pid[1] != "Finished":
+				if pid[1] != "Finished" and pid[1] != "Killed":
 					os.kill(pid[0].pid, signal.SIGKILL)
 
 def verify_config(mode, configList):
@@ -19,13 +19,17 @@ def verify_config(mode, configList):
 	all required parameters
 	"""
 	i = 0
+	totalinstances = 0
 	while i < len(configList):
 		j = i + 1
+		totalinstances = totalinstances + configList[i][2]
 		while j < len(configList):
 			if configList[i][0] == configList[j][0]:
 				return errors.error_repeated_names(mode)
 			j += 1
 		i += 1
+	if totalinstances > 400:
+		errors.error_instances(mode)
 	for config in configList:
 		if len(config) != 15:
 			return errors.error_config_len(mode)
@@ -49,7 +53,7 @@ def verify_config(mode, configList):
 			"starttime/stoptime/restartretries")
 		elif (isinstance(config[8], str) != True or (config[8] != "TERM" and
 		config[8] != "QUIT" and config[8] != "INT" and config[8] != "KILL"
-		and config[8] != "TSTP" and config[8] != "HUP" and config[8] != "STOP" )):
+		and config[8] != "TSTP" and config[8] != "STOP" )):
 			return errors.error_config(mode, config[0], "quitsig")
 		elif isinstance(config[9], str) != True and isinstance(config[9], list) != True:
 			return errors.error_config(mode, config[0], "exitcodes")
